requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Deprecated==1.2.15
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
frozenlist==1.5.0
future==1.0.0
idna==3.10
line-bot-sdk==3.14.2
multidict==6.1.0
pdf2image==1.17.0
pillow==11.0.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.2
